refactor: utils dir for directed entities
