+++ otary/geometry/discrete/linear/directed/utils/direction.py
